Clean code 🧹
