+++ apps/ledger-live-desktop/src/renderer/screens/market/MarketList.tsx
@@ -1,5 +1,4 @@
 import React, { useCallback, memo } from "react";
-import styled, { DefaultTheme, StyledComponent } from "styled-components";
 import { Flex, Text, Icon } from "@ledgerhq/react-ui";
 import { TFunction } from "i18next";
 import { Trans, useTranslation } from "react-i18next";
@@ -7,7 +6,6 @@ import { FixedSizeList as List } from "react-window";
 import InfiniteLoader from "react-window-infinite-loader";
 import AutoSizer from "react-virtualized-auto-sizer";
 import MarketRowItem from "./MarketRowItem";
-import LoadingPlaceholder from "../../components/LoadingPlaceholder";
 import { Button } from ".";
 import { useSelector, useDispatch } from "react-redux";
 import { localeSelector } from "~/renderer/reducers/settings";
@@ -17,7 +15,6 @@ import { useRampCatalog } from "@ledgerhq/live-common/platform/providers/RampCat
 import Image from "~/renderer/components/Image";
 import NoResultsFound from "~/renderer/images/no-results-found.png";
 import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
-import { FlexProps } from "styled-system";
 import {
   CurrencyData,
   MarketListRequestParams,
@@ -28,155 +25,8 @@ import { useFetchCurrencyFrom } from "@ledgerhq/live-common/exchange/swap/hooks/
 import { useMarketData as useMarketDataHook } from "@ledgerhq/live-common/market/v2/useMarketDataProvider";
 import { marketSelector } from "~/renderer/reducers/market";
 import { setMarketOptions } from "~/renderer/actions/market";
-
-export const TableCellBase: StyledComponent<"div", DefaultTheme, FlexProps> = styled(Flex).attrs({
-  alignItems: "center",
-})<{ disabled?: boolean }>`
-  padding-left: 5px;
-  padding-right: 5px;
-  cursor: ${p => (p.disabled ? "default" : "pointer")};
-`;
-
-export const TableCell = ({
-  loading,
-  children,
-  ...props
-}: {
-  disabled?: boolean;
-  loading?: boolean;
-  onClick?: (e: React.MouseEvent<HTMLDivElement>) => void;
-  children?: React.ReactNode;
-} & React.ComponentProps<typeof TableCellBase>) => (
-  <TableCellBase {...props}>
-    {loading ? <LoadingPlaceholder style={{ borderRadius: 50, overflow: "hidden" }} /> : children}
-  </TableCellBase>
-);
-
-const ChevronContainer = styled(Flex).attrs({ m: 1 })<{
-  show: boolean;
-  orderDirection?: string | undefined;
-}>`
-  opacity: ${p => (p.show ? 1 : 0)};
-  svg {
-    transform: rotate(
-      ${p => (p.orderDirection ? (p.orderDirection === "asc" ? "180deg" : "0deg") : "90deg")}
-    );
-    transition: transform 0.3s ease-out;
-  }
-`;
-
-export const miniChartThreshold = 900;
-export const miniMarketCapThreshold = 1100;
-
-export const SortTableCell = ({
-  onClick,
-  orderByKey,
-  orderBy,
-  order,
-  children,
-  ...props
-}: {
-  loading?: boolean;
-  onClick?: (key: string) => void;
-  orderByKey: string;
-  orderBy?: string | undefined;
-  order?: string | undefined;
-  children?: React.ReactNode;
-}) => (
-  <TableCellBase onClick={() => !!onClick && onClick(orderByKey)} {...props}>
-    {children}
-    <ChevronContainer m={2} show={orderBy === orderByKey} orderDirection={order}>
-      <Icon name="ChevronBottom" size={10} />
-    </ChevronContainer>
-  </TableCellBase>
-);
-
-const listItemHeight = 73;
-
-export const TableRow: StyledComponent<
-  "div",
-  DefaultTheme,
-  FlexProps & { header?: boolean; disabled?: boolean }
-> = styled(Flex).attrs({
-  flexDirection: "row",
-  alignItems: "stretch",
-  justifyContent: "flex-start",
-  height: listItemHeight,
-  py: 3,
-})<{ header?: boolean; disabled?: boolean }>`
-  line-height: 16px;
-  ${p =>
-    p.header
-      ? `
-    color: ${p.theme.colors.neutral.c80};
-    font-size: 12px;
-    padding-right: 12px;
-  `
-      : `
-    color: ${p.theme.colors.neutral.c100};
-    font-size: 13px;
-    border-bottom: 1px solid ${p.theme.colors.neutral.c40};
-
-    :hover {
-      background:  ${p.theme.colors.neutral.c20};
-    }
-    :active {
-      background: ${p.theme.colors.neutral.c30};
-    }
-  `}
-
-  cursor: ${p => (p.disabled ? "default" : "pointer")};
-
-  ${TableCellBase}:nth-child(1) {
-    flex: 0 0 40px;
-    justify-content: flex-start;
-    padding-left: 5px;
-  }
-  ${TableCellBase}:nth-child(2) {
-    flex: 1 0 230px;
-    justify-content: flex-start;
-  }
-  ${TableCellBase}:nth-child(3) {
-    flex: 1 0 80px;
-    justify-content: flex-end;
-  }
-  ${TableCellBase}:nth-child(4) {
-    flex: 1 0 30px;
-    justify-content: flex-end;
-  }
-  ${TableCellBase}:nth-child(5) {
-    flex: 1 0 90px;
-    justify-content: flex-end;
-  }
-  ${TableCellBase}:nth-child(6) {
-    flex: 1 0 70px;
-    justify-content: flex-end;
-  }
-
-  ${TableCellBase}:nth-child(7) {
-    flex: 0 0 40px;
-    justify-content: flex-end;
-    padding-right: 5px;
-    svg {
-      fill: currentColor;
-    }
-  }
-
-  @media (max-width: ${miniChartThreshold}px) {
-    ${TableCellBase}:nth-child(6) {
-      display: none;
-    }
-  }
-
-  @media (max-width: ${miniMarketCapThreshold}px) {
-    ${TableCellBase}:nth-child(3) {
-      flex: inherit;
-    }
-    ${TableCellBase}:nth-child(1), ${TableCellBase}:nth-child(5) {
-      display: none;
-    }
-  }
-`;
+import { SortTableCell } from "./components/SortTableCell";
+import { TableCell, TableRow, listItemHeight } from "./components/Table";
 
 const NoCryptoPlaceholder = ({
   requestParams,
